style: playlist 반응형 레이아웃 및 스타일링 완료
