+++ syncfit/src/pages/PlayList.jsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import RecommendedPlaylist from '../components/RecommendedPlaylist'
+import WithoutFooterLayout from '../components/Layout/WithoutFooterLayout'
 
 const PlayList = () => {
   return (
-    <>
+    <WithoutFooterLayout>
       <RecommendedPlaylist />
-    </>
+    </WithoutFooterLayout>
   )
 }
